diff: improve diff readability (#17)
* diff: improve diff readability

By implementing a custom diff reporter, this will improve diff
readability. It's still not perfect, but it's not extremely ugly like
before.

* fix go install

* diff: reporter outputs empty str with no diffs

Instead of messing with a boolean return value from the multiline diff
helper, just return empty string. That way presence of a diff can be
checked by the string being empty.