feat: add ability to pass config path (#30)
* feat: add ability to pass config path

This PR adds a feature I originally intended for the first version but
forgot about until the opened issue. The tool now accepts a new flag
`-conf` and accepts a relative or absolute path to a config file. If the
flag is not present, it will fall back to the default config.

* docs: Add usage for conf flag

* docs: fix indentation setting