fix: restore functionality of -in flag
Originally I had changed the read from stdin to be specified by a single
- argument or /dev/stdin by suggestion from #26. In that PR I had
removed the -in flag. After sleeping on it I realized I might as well
keep that flag, since it might intuitively make sense to more people,
makes more sense on Windows, and maintains backwards compatibility (even
though that is currently not a promise of the project, might as well do
it if it's easy).