yamlfmt: added and implemented functional features (#43)
* yamlfmt: added and implemented functional features

While adding functionality for small edge cases or formatting
differences, I found that the Formatter code was becoming harder to
reason about. It would grow linearly with the number of little
formatting things I had to do. I have adjusted the pattern to instead be
a collection of functional features that can be added at-will.

Created a new `Feature` type that has a Before and After action. Also
added the FeatureList type, with code to apply the before or after
actions in order on byte slice input.

I converted all existing candidates in the formatter code to use this
functional feature pattern instead.

* hotfix: make hotfix package create feature

* yamlfmt: wrap feature apply errors

This wraps feature apply errors with the name and type of action.

* basic: change feature configuration pattern

Instead of a receiver method that configures the features in place on
the formatter, make it so the features are returned from a function that
reads the config. Avoids deliberate mutability.