feat: global linesep character, handle diffs
Using the new `multilinediff` there needed to be two changes made:
* Needed to handle diffs using the count of differences found, since the
  new reporter always reports something
* Needed to recongize a line separation character at the engine level

This PR fixes the diff handling, which was relatively trivial. However,
an oversight since the beginning is that diffs never took into account
the line ending being used, and assumed `\n`. This needed to be
rectified, and lead to something of a re-architecture that adds
`line_ending` as a setting that acts as a global override. This required
simplifying the `Factory` interface a little bit, in a way that sort of
simplifies it. While refactoring anyway, I also decided to simplify the
line break logic, so I didn't have to repeat the same constant check
everywhere.