paths: do not include directories in formatting
If a directory was named such that it matched the include pattern (in
most circumstances, if it was named ending in `.yml`/`.yaml`) then
yamlfmt would try to read it and format it which would fail because it's
a directory. This PR adds a stat check to ignore all directories. Also
adds tests for this package which did not previously exist.

Signed-off-by: braydonk <braydonk@google.com>