yamlfmt: created path collectors
Created a new PathCollector interface that can accept different types of
patterns and collect paths to format in different ways. Use the
interface in Engine, and create the right collector in Command based on
the config. The new default will be the FilepathCollector, which won't
actually collect any paths without being told, and adding
`doublestar: true` to the config or `-dstar` as a flag will restore old
behaviour.