engine: overhauled with interface and better code
While addressing the new quiet mode, the overall weakness of the engine
code was very exposed. I decided to make the effort to improve it,
creating an interface and a far more sane code path overall that will be
more maintainable in the long term. This also opens up the opportunity
for me to create a new alternate engine that will do all of this
formatting and processing in parallel for users with larger codebases.