engine: overhauled with interface and better code (#90)
* engine: overhauled with interface and better code

While addressing the new quiet mode, the overall weakness of the engine
code was very exposed. I decided to make the effort to improve it,
creating an interface and a far more sane code path overall that will be
more maintainable in the long term. This also opens up the opportunity
for me to create a new alternate engine that will do all of this
formatting and processing in parallel for users with larger codebases.

* ci: change to go 1.20

* main: add EngineOutput type

Adding the EngineOutput type for more consistent output messaging rather
than constructing strings in place in the engine.

* main: change FileDiffs to a map to avoid duplicates

* tempfile: Added tempfile package for tests

Abandoning engine tests right now cause they'll take too long but I'll
include in the future integration testing plans.

* docs: Added quiet flag, restructured config docs

* engine: improve dry run output message