yamlfmt: created content analyzer (#106)
* yamlfmt: created content analyzer

In some use cases the provided patterns of exclusion aren't enough, for
example when something generates a lot of yaml files in weird places, or
when single files should be ignored out of large directories. This PR
adds the following:
* The concept of and ability to read metadata, with the first type of
  metadata being "ignore"
* A new ContentAnalyzer interface and a BasicContentAnalyzer which will accept
  an array of regex pattern strings. This will first read the metadata
  from the content of each file to find the ignore metadata, then will
  match the content to the regex patterns provided to determine which
  patterns will be excluded.

* yamlfmt: improve metadata errors

Change metadata errors into a wrapped error struct so that the line
number and path can be included in the resulting error. Make metadata
errors non-fatal to `yamlfmt` as a whole. Also add the docs in this
commit.