yamlfmt: fix standard mode exclude bug (#124)
Due to a misunderstanding on my part, I swapped standard filepath
collection mode to calculate exclusions using absolute paths. This
worked because of the way tests were written, but it does not work in
most real world scenarios and it slipped by. I will be able to catch
this better when I implement proper integration testing.

This PR removes the absolute paths from the calculation, and adds a doc
entry to recommend that paths are always specified relative to the
command working directory.