Debug logging
This PR adds debug logging. Debug logging can be enabled through flags
to the yamlfmt command, and will tell you in very noisy detail about
things that are going on. Debug groups are separated into codes, so that
only certain debug logs can be enabled if you are debugging a specific
problem.

I also fixed a bug I introduced in the last config PR with the `-conf`
flag. It is fixed before ever going out in a release, so it will only
affect someone who installed from that exact commit before I fix it
here.