engine: return nil from dry run when no changes
The `command` output actually relies on `nil` being return when there
are no changes, and I forgot to add that behaviour to dry run which made
there be no message printed out for confirmation.

Also decided to make a slight change to the confirmation message so that
it outputs nothing when the `-quiet` flag is enabled.

Signed-off-by: braydonk <braydonk@google.com>