Vendor github.com/braydonk/yaml (#245)
* Vendor github.com/braydonk/yaml

The yaml.v3 package upstream is now archived. This means there is no hope of my patches ever being committed upstream. As such, there is
nothing to gain by maintaining this library in a separate repo.
It will be easier for people to contribute to if it's in this repo.
I will keep github.com/braydonk/yaml alive purely for the select few who pull
the package for the couple of patches that upstream never merged.

* fix test workflows cause the yaml package fails everything