pkg/yaml: correct alias keys (#247)
* pkg/yaml: correct alias keys

As outlined in #247, the yaml library did not initially handle this edge
case of using an alias node as a key properly. This PR adds
functionality to correct that. In case there is some other edge case
where this breaks, the `disable_alias_key_correction` formatter config
option was added.

* forgot to fix this comment

* forgot to enable the setting for this test