formatters/basic: fix just comments edge case (#259)
When `yaml.v3` parses a document that's just comments, it ends up
parsing no actual nodes and thus loses the comment information. This
will now be treated as an edge case and return the original content, in
essence not modifying it if the yaml parsed successfully but no nodes are
parsed out.