Remove trailing whitespace from merge aliases. (#260)
* fix(yaml): Remove trailing whitespace from merge aliases.

PR #247 introduced a regression with YAML documents such as this:

```yaml
config: &shared
  key: value

instance:
  <<: *shared
```

Due to the "correct alias keys" setting, a extra space was emitted after
`*shared`. This created trailing whitespace which many tools (rightly) complain
about.

This change adds an extra condition to only emit the space in the "simple key"
context. I'm not enough of a YAML expert to be sure this is the right thing to
do, but it fixes my problem without failing any of the existing tests.

* test(formattest): Print a diff between the expected and actual result.