pre-commit: stop specifying types in hook (#274)
* pre-commit: stop specifying types in hook

This PR updates the pre-commit hook to not force `yaml` files. `yamlfmt`
is designed to accept all files in a working directory and filter
through what files are accepted based on configuration (i.e. it's
perfectly valid for users to configure `yamlfmt` to accept files that
wouldn't match pre-commit's `yaml` type).

* also update hook version cause I'm gonna bump for this