pre-commit: don't pass filenames (#277)
By default, hooks will pass filenames to yamlfmt. On purpose, yamlfmt
will handle this by accepting each file argument and passing it through
the formatter with no guards. This is not a good default experience.
This PR changes the hook to not pass filenames, and the default `entry`
to be `yamlfmt .`. This should match the experience I thought the hook
would be giving all along.