New Formatter KYAML
This PR adds the first new formatter to yamlfmt, `kyaml`. This will read
any YAML document and use the official KYAML encoder to format it as
KYAML.

I found the usability a bit tricky for an alternate formatter, so since
this is the first one to exist I made some changes to how that worked.
Namely, I added a simple `-kyaml` flag that will override the configured
formatter and use `kyaml` formatter instead, which allows users to
quickly try it ad-hoc without hoops.