Force single or double quotes
This PR adds `force_quote_style` to the basic formatter. This will allow
users to for `'` or `"` around all strings. Along with this I ended up
doing some refactoring for config validation to make it so invalid
values for some enum options will actually fail instead of just silently
doing nothing. I also hit the last straw with the tests in this package
so I totally redid them. I could have done it in another PR so this one
isn't too big, but I'm the only person who works on this so as the only
person inconvenienced I say it's fine. :)