Fix filepath collector panic
Currently, the situation is in the rare case when the filepath package
can't walk, yamlfmt will then call a function on a nil fileinfo which
will panic. This PR adds handling for the error from filepath walking
which will print the errors for the paths it can't read and format what
it is capable of formatting.